Generate short key the is not in the db
Before, the short key was only a generated random string of length 4.
There was no check if that key is already in the database.

Now, a special function is added that checks if the generated short key
is already in the database, before it is used to save a new long URL.

But, this is not concurrency proof. If the short key is added to the
table between the checking of the short key existence and the adding of
the new item it is still possible that the new key generation fails.