Migrate script for production data, no sort key
Before, the production table was using a compound primary key consisting
of the hash key `short` and the sort key `long_url`. This lead to the
database not throwing duplication errors when the same `short` was
used with different `long_url`.

A script was created to help the migration, although it was partially
executed manually and original steps had to be commented out. I am only
adding this script for the purpose of documentation.

Both, the script and the changes in the default database connection
definition will be remove in the next commit, because they wont be
needed anymore.