Handle duplicate short prevention in db
Before, the prevention of duplicate `short` keys was handled on the app
level. That implementation was prone to concurrency issues.

Now, the duplication is handled on the db level which is safe against
the aforementioned concurrency issues.

The concurrency issues where unlikely before, because it would require
the generation of the same random key with in the same time frame and
perfectly offset processes. But it was possible. The change was
introduced to handle the duplication issue in a cleaner and db native
way.